requirements.txt
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
openai>=1.0.0
pinecone-client>=3.0.0
streamlit>=1.33.0
sentence-transformers>=3.0.0
